Add files via upload
In this project, I have introduced the Bayesian optimization technique to improve the accuracy of the machine learning (MNB, RF, DT) and deep learning model (CNN). For this experiment, a spam or ham-labeled SMS message collection dataset from Kaggle was used. The  experimental results show that Bayesian optimized CNN got the highest accuracy of 98.5% compared to other implemented models.
I have also showed the comparison of results with and without preprocessing. And also the results of with and without using Bayesian Optimization Technique.